Switched GetArea() to a double method. Circle class added.

+++ prepare/Learning05/Rectangle.cs
@@ -1,13 +1,13 @@
 public class Rectangle : Shape
 {
-    private int _length;
-    private int _width;
-    public Rectangle(string color, int length, int width) : base(color)
+    private double _length;
+    private double _width;
+    public Rectangle(string color, double length, double width) : base(color)
     {
         _length = length;
         _width = width;
     }
-    public override int GetArea()
+    public override double GetArea()
     {
         return _length * _width;
     }
